+++ app/src/main/java/com/theandroiddev/mywins/data/successes/model/SuccessEntity.kt
@@ -1,8 +1,5 @@
-package com.theandroiddev.mywins.data.model
+package com.theandroiddev.mywins.data.successes.model
 
-import android.arch.persistence.room.ColumnInfo
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.PrimaryKey
 import com.theandroiddev.mywins.domain.service.successes.SuccessesServiceModel
 import com.theandroiddev.mywins.local.model.LocalSuccess
 import com.theandroiddev.mywins.utils.Constants.Companion.Category
