+++ app/src/main/java/com/theandroiddev/mywins/data/successes/source/SuccessesLocalDataSourceImpl.kt
@@ -1,12 +1,10 @@
-package com.theandroiddev.mywins.data.source
+package com.theandroiddev.mywins.data.successes.source
 
 import android.arch.persistence.db.SimpleSQLiteQuery
-import android.arch.persistence.db.SupportSQLiteQuery
-import com.github.ajalt.timberkt.e
 import com.theandroiddev.mywins.local.dao.SuccessDao
-import com.theandroiddev.mywins.data.model.SuccessEntity
-import com.theandroiddev.mywins.data.model.toLocal
-import com.theandroiddev.mywins.data.repository.SuccessesLocalDataSource
+import com.theandroiddev.mywins.data.successes.model.SuccessEntity
+import com.theandroiddev.mywins.data.successes.model.toLocal
+import com.theandroiddev.mywins.data.successes.repository.SuccessesLocalDataSource
 import com.theandroiddev.mywins.local.model.toEntity
 import com.theandroiddev.mywins.utils.Constants
 import com.theandroiddev.mywins.utils.Constants.Companion.dummyCategories
